ENH: Handle sequences with ambiguous nucleotides
Ambiguous nucleotides will only be used when counting totals (total number of dinucleotides, total number of amino acids) but will not be included otherwise; e.g. if a codon is not in the codon table, it will not be counted as a particular amino acid & there will not be features for dinucleotide bias of pairs including ambiguous nucleotides.