BUG: restore ability to run DTRA code
* Fixes #74 ; this is currenly a work in progress that
I'd like Adam to take over/review/fix up. This branch
seems to allow for the DTRA workflow to be executed
and produce plots again after Abhishek broke it
in !49 and didn't fix it before moving on.

* The fix includes starting to remove the use of
the silly `Unnamed` column that duplicates the index
per #70 --it may make sense to split that off into
a separate MR with testing.

* The `csv_conversion` function is now used with
a temporary file so that we adjust the input CSV
data for binary classification/comparisons programmatically
rather than having to adopt a bunch of extra columns
in the CSV we store in the repo.

* A bunch of code indentation errors were introduced
by Abhishek--I've tried to fix some of them, but there
are likely still issues, quite a mess!

* I've backed out some of the `pearson_r` k-mer effect
tolerance changes that Abhishek made because those
need to be discussed/reviewed/designed first...