CI: Python 3.11 testing
* This branch deals with a small part of #77. As noted
there, we've been stuck testing a single, fairly old,
version of Python for 1.5+ years now, and there hasn't
been much proactive activity to improve our portability
testing, so I'm trying to nudge this forward a bit here.
It would be great if the postdocs on the team could start
taking ownership of these types of things.

* Adds Python 3.11 testing to the CI, and retains
the old Python 3.10 testing, by using a `matrix` directive,
so that we have improved portability testing.

* There are some drawbacks to how this new approach to CI
works relative to the old way, but none of them are remotely
as important as testing more versions of Python.

* Summarizing some of the changes/side effects of the new
`matrix` approach to CI here: to avoid re-pulling the separate
docker images, the steps get consolidated to a single "job"
and single "stage" which is a bit annoying in the pipeline
visualization UI (though at least 3.10 and 3.11 jobs to get
separated if you click on the pipeline); the linting steps
are repeated for both Python versions (not much benefit to that).

* I did intentionally change the pins on a few dependency versions
where it was either helpful or just didn't make sense to pin
to an old version (i.e., `pytest` should always be latest stable
release really).

* I've removed some of the restrictions on where the CI pipeline
runs--that was helpful for debugging this branch by allowing
the CI pipeline to always run when you push up to GitLab, even
when not in an active MR. This should be fine as long as folks
use the appropriate `[ci skip]` style message in their commits
when they don't need to flush CI.