MAINT, CI: ruff fixes, bump
* Following the merge of !93 the CI failed
on `main` because I had accidentally deactivated
a bunch of `ruff` rules (which Aaron recently
restored/reativated). So, this branch includes
the removal of two extraneous f-string prefixes
and one unused variable that slipped through.

* After fixing that, also bump `ruff` to the latest
stable release `0.8.1` while I'm in the neighborhood
(linting continues to pass, consistent with our
selection of an explicit set of rules that is stable
between releases).