BUG: remove averaging multiple calibration curves
The method of averaging calibration curves was not valid as it did not
take into account that classifiers do not always have predictions across
the full range of [0,1]. This makes it unclear the best approach to
average calibration curves which may have different domains.

Calibration plots summarizing a group of one type of classifiers have
been reverted to the previous behavior of averaging the predictions
before generating the plot.