MAINT, DOC: MR 100 revisions
* Add a `ruff` rule to detect unused function
arguments, ignore violations of this rule unrelated
to the current MR, and fix the one violation of this
rule that is related to this MR. The latter fix
also addressed a reviewer comment related to `build_tables`
accessing a module-global variable instead of having
it passed in.

* Fix the `--kmer-range` help/documentation to clarify
that a single kmer length value is now accepted, and
add error handling for the case where the provided
lower bound exceeds the upper bound kmer length.