ENH: support Nick H. data request
* Fixes #128

* Adjust `drop_unshared_kmers()` function to support a new
argument, `consensus_req`, that allows the user to specify
how many viral sequence records in the design matrix must
contain a given kmer for it to be retained in our workflow.
`consensus_req` will default to `2` to remain backwards
compatible with our usual workflow settings. This is to
accommodate Nick's request of a design matrix that retains
only kmers shared by `4` viral sequence records.

* The above functionality is user-exposed in the workflow via
a new CLI flag, the `--kmer-consensus` integer value. Propagating
the capability throughout the workflow requires a fair number
of shims.

* Add draft regression test, `test_drop_unshared_kmers()`, to
enforce above behavior internally.

* In `feature_selection_rfc()`, I believe it is necessary
to save `X_train` in a broader set of cases so that we
can accommodate these requests when feature importance
filtering is not requested, so `X_train` is now also
written to disk when feature importance filtering
is not requested.