TST, BUG: PR 37 revisions
* `test_grab_features_kmer_maps()` was adjusted to account
for inclusion of both AA:AA and PC:AA `KmerData` instances.
It wasn't clear to me if we should only keep the PC:AA mappings
or all of them, so I've added a detailed comment to this test
about that.

* `test_save_load_all_kmer_info()` was adjusted to allow an extra
column in the expected data structure for `all_kmer_info_df`.

* Tentative shims to `get_kmers()` that allow the full testsuite to
  pass.