MAINT: limit estimator concurrency
* Related to the discussion in gh-39, where some DTRA workflow
incantations were taking on the order of 90+ hours to run on
a Linux box (i.e., `gp160.lanl.gov`) that has many cores/threads,
which is far too slow to be practical.

* Adam had originally proposed to limit concurrency by hard coding
the number of threads to use for XGBoost and LGBM estimators to 16,
but there were a few issues with that solution:

1) 16 is a fairly arbitrary number of threads in the grand scheme of things, and may
not scale well on many machines, so I've replaced that approach
with automatic counting of the number of physical cores and usage
of that number of threads for XGB and LGBM in the DTRA workflow. This
required addition of a new dependency, `psutil`.

2) Adam proposed using `thread` to control the concurrency for
`XGBClassifier` at the same time as using `n_jobs`, which is not
only confusing, but also deprecated by the xgboost team 8 years ago:
https://github.com/dmlc/xgboost/releases/tag/v0.7
I don't know why this was suggested, and I didn't receive a response
from Adam, but it seems sensible to try the modern `n_jobs` rather
than both options at the same time, especially when one of them is
deprecated.

3) Adam proposed using `num_threads` and `n_jobs` at the same time
for controlling the concurrency of LGBM. For many of the same reasons,
this is confusing to me without a clear justification, the latest docs
only show `n_jobs` for `LGBMClassifier` and that is the typical
`sklearn` estimator API, so again let's try to use `n_jobs` here.

* Before this is merged, it should be confirmed that one of the
slow/problem incantations runs in a reasonable amount of time
on both an ARM Mac and on gp160.